Add a flag to switch between LAPACK and decompose() + small changes.

+++ Addons/testExoticQuadrature.hpp
@@ -44,7 +44,7 @@ inline bool testNodeAndWeightSizes() {
     std::vector<double> ref_weights(nref), ref_points(nref);
     TasGrid::OneDimensionalNodes::getGaussLegendre(nref, ref_weights, ref_points);
     // n = 1, 2, 5
-    std::vector<int> n_vec = {1, 2, 5};
+    std::vector<size_t> n_vec = {1, 2, 5};
     for (auto n : n_vec) {
         TasGrid::getGaussNodesAndWeights<true>(n, ref_points, ref_weights, nodes_cache, weights_cache);
         if (nodes_cache.size() != n) {
